[Matthew] - Merged John's attack animation movements to the branch
