+++ PreproductionProject/Assets/Prefabs/GameDependencies.prefab
@@ -10,7 +10,7 @@ GameObject:
   m_Component:
   - component: {fileID: 26918801}
   - component: {fileID: 26918804}
-  - component: {fileID: 26918802}
+  - component: {fileID: 2807173402665339149}
   m_Layer: 0
   m_Name: Collider
   m_TagString: Untagged
@@ -45,7 +45,7 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 1, y: 1, z: 1}
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &26918802
+--- !u!114 &2807173402665339149
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -54,7 +54,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 26918800}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3d769b0a4975c4e4abcda1dea803c1f7, type: 3}
+  m_Script: {fileID: 11500000, guid: bb6844229f3ab204aaf4807e809b8051, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &98351112
@@ -3102,14 +3102,14 @@ GameObject:
   - component: {fileID: 6226596333264784430}
   - component: {fileID: 1615343467}
   - component: {fileID: 1038329704}
-  - component: {fileID: 3838162160385988825}
+  - component: {fileID: 7443772372247003620}
+  - component: {fileID: 5232862844244932280}
   - component: {fileID: 6226596335058991072}
   - component: {fileID: 6226596335058991075}
   - component: {fileID: 6226596335058991074}
   - component: {fileID: 6226596335058991085}
   - component: {fileID: 6226596335058991084}
   - component: {fileID: 6226596335058991087}
-  - component: {fileID: 2333746424419587477}
   m_Layer: 0
   m_Name: NewPlayerModel
   m_TagString: PlayerTag
@@ -3175,7 +3175,7 @@ MonoBehaviour:
   ui_offset: {x: 0, y: 0}
   targets: []
   range: 20
---- !u!114 &3838162160385988825
+--- !u!114 &7443772372247003620
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -3184,9 +3184,24 @@ MonoBehaviour:
   m_GameObject: {fileID: 6226596335058991086}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 12b796f0e8a7fdb49a7112917a1323a7, type: 3}
+  m_Script: {fileID: 11500000, guid: 3d769b0a4975c4e4abcda1dea803c1f7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  attackCollider: {fileID: 26918804}
+--- !u!114 &5232862844244932280
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6226596335058991086}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e176d04b930b03e4e99b04bb89b3d29a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  speed: 5
+  doMove: 0
 --- !u!114 &6226596335058991072
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -3335,19 +3350,6 @@ CapsuleCollider:
   m_Height: 1.8
   m_Direction: 1
   m_Center: {x: 0, y: 0.9, z: 0}
---- !u!114 &2333746424419587477
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6226596335058991086}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 23176b5c768a9f74abec85b508774c49, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  attackCollider: {fileID: 26918804}
 --- !u!1 &6226596335167636599
 GameObject:
   m_ObjectHideFlags: 0
