+++ PreproductionProject/Assets/Scripts/John_Scripts/AttackManager.cs
@@ -1,27 +1,73 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.UI;
+using System;
 
 public class AttackManager : MonoBehaviour
 {
-    NewPlayerScript playerScript;
-    private float attackDmg;
+    public Action OnAttackStart, OnAttackStop;
+
+    [SerializeField]
+    Collider attackCollider;
     
-    private void Awake()
+
+    Animator animator;
+    Rigidbody rigidBody;
+    bool isAttacking;
+
+    int attackIndex = 0;                 // Determines which animation will play
+    bool canClick;                  // Locks ability to click during animation event
+    private const int totalAttacks = 3;
+
+    float delayAttack = 0.0f;
+
+    void Start()
     {
-        playerScript = FindObjectOfType<NewPlayerScript>();
-        attackDmg = playerScript.AttckDamage;
+        animator = GetComponent<Animator>();
+        rigidBody = GetComponent<Rigidbody>();
+        attackCollider.gameObject.SetActive(false);
+
+        attackIndex = 0;             // numbers of clicks
+        canClick = true;
     }
-    private void OnTriggerEnter(Collider other)
+
+    // Update is called once per frame
+    void Update()
     {
-        if (other.gameObject.CompareTag("EnemyTag"))
+        if(Input.GetMouseButtonDown(0) && delayAttack < Time.time && Time.timeScale != 0)
         {
-            BaseEnemyScript enemy = other.GetComponent<BaseEnemyScript>();
-            enemy.TakeDamage(attackDmg);
-            playerScript.IncrementHitPoint();
-        }
+            delayAttack = Time.time + 0.8f;
+
+            Attack();
+        }      
+    }
+
+    void Attack()
+    {
+        isAttacking = true;
+        string attackTrigger = "Attack" + (attackIndex+1).ToString();
+
+        animator.SetTrigger(attackTrigger);
+
+        if (attackIndex < totalAttacks-1)
+            attackIndex++;
+        else
+            attackIndex = 0;
+
+        OnAttackStart?.Invoke();
+    }
+
+    void AttackStart()
+    {
+        attackCollider.gameObject.SetActive(true);
+    }
+
+    void AttackStop()
+    {
+        attackCollider.gameObject.SetActive(false);
+
+        isAttacking = false;
+
+        OnAttackStop?.Invoke();
     }
 }
-
-
