+++ PreproductionProject/Assets/Scripts/John_Scripts/Weapon.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 90793bba2f17a3a44a66db44a63c3ad0
+guid: bb6844229f3ab204aaf4807e809b8051
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
